# api/openapi.yaml
openapi: 3.1.0
info:
  title: Hefesto API
  version: 0.4.0
  summary: Calisthenics training logger and skill map
  description: |
    The source of truth for the Hefesto API and for the generated Swift client.
    DTOs are never hand-written; change this file first.

    ## Conventions

    - JSON bodies use `snake_case`. Instants are RFC 3339 in UTC; `local_date`
      is the athlete's calendar day in the session's own time zone.
    - Every row a client creates carries a **client-generated UUIDv7 `id`**.
      This is the same contract the delta sync (Phase 4) uses, so an object
      created over REST and one created offline are indistinguishable.
    - `updated_at` on a write is the client's clock; the server records its own
      `server_updated_at` beside it.
    - **One code path for sets.** A set of 8 pull-ups is one set entry with one
      element; a combo is one set entry with N ordered elements. There is no
      endpoint that writes an element on its own.
    - Deleting is soft: the row keeps its id with `deleted_at` set, so other
      devices learn about the deletion when they sync.

    ## Errors

    Every error is `application/problem+json` (RFC 9457). Clients switch on
    `type`, never on `detail`. `errors` maps a JSON pointer into the request
    body to a message, for validation failures.

    | `type` (under `https://hefesto.fit/problems/`) | Status | Meaning |
    |---|---|---|
...
    | `unauthorized` | 401 | Missing, malformed or expired access token. |
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/sessions/{sessionId}/complete:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: completeSession
      summary: Complete a session and evaluate unlocks
      description: |
        Marks the session completed, then evaluates every skill level against
        the athlete's completed sessions and returns what changed: levels
        unlocked (for the celebration), levels newly available, XP awarded
        and the streak.

        Idempotent: completing an already completed session changes nothing
        and returns the unlocks that session caused, with
        `already_completed: true`. Rest-day sessions keep the streak but earn
        no XP and unlock nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionComplete"
      responses:
        "200":
          description: Completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionResult"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"

  # ---------------------------------------------------------------- skills
  /v1/skills:
    get:
      tags: [skills]
      operationId: getSkillGraph
      summary: The whole skill graph
      description: |
        Every skill that is not retired, with its levels, their exercises and
        unlock criteria, and every edge between levels. Map positions are
        authored, not computed. The response carries an `ETag` keyed on the
        content version; a matching `If-None-Match` returns 304.
      parameters:
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        "200":
          description: The graph.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGraph"
        "304":
          description: The graph has not changed.
        "401":
          $ref: "#/components/responses/Problem"
  /v1/skills/{slug}:
    get:
      tags: [skills]
      operationId: getSkill
      summary: One skill, with its injury notes
      description: Injury notes are educational, never medical advice; each carries its disclaimer, and the payload carries the disclaimer text.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Slug"
      responses:
        "200":
          description: The skill.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDetail"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/skill-map:
    get:
      tags: [skills]
      operationId: getMySkillMap
      summary: The graph and my standing on every level, in one payload
      responses:
        "200":
          description: The map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillMap"
        "401":
          $ref: "#/components/responses/Problem"
...
  /v1/me/progress:
    get:
      tags: [skills]
      operationId: getMyProgress
      summary: My XP and streak
      responses:
        "200":
          description: Progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
        "401":
          $ref: "#/components/responses/Problem"

  # ------------------------------------------------------------------ sync
...
  /v1/media/uploads:
    post:
      tags: [media]
      operationId: createUpload
      summary: Reserve a media asset and get a presigned upload URL
      description: |
        Creates a `pending` asset and returns a short-lived presigned `PUT`
        URL. Upload the bytes directly to object storage with exactly the
        returned headers, then call `POST /v1/media/{mediaId}/complete`.
        Images only for now (ADR 0002 §4); the schema is ready for video.
        Repeating the call for a pending asset returns a fresh URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadRequest"
      responses:
        "201":
          description: Reserved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUpload"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/media/{mediaId}/complete:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    post:
      tags: [media]
      operationId: completeUpload
      summary: Confirm an upload
      description: |
        Checks the uploaded object against what was reserved (size and
        content type) and marks the asset `ready`. An object that does not
        match is removed and the asset marked `failed`. Idempotent.
      responses:
        "200":
          description: Ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      tags: [media]
      operationId: getMedia
      summary: A media asset, with a short-lived download URL when ready
      responses:
        "200":
          description: The asset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [media]
      operationId: deleteMedia
      summary: Delete a media asset and its stored object
      description: The asset is detached from any set element it was attached to.
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    Problem:
      description: An error occurred.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Id:
      description: A UUID. Rows a client creates carry a client-generated UUIDv7.
      type: string
      format: uuid
    Slug:
      type: string
      pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
      maxLength: 80
    Timestamp:
      type: string
      format: date-time
    NullableTimestamp:
      type: [string, "null"]
      format: date-time
    TimeZone:
      description: An IANA time-zone name.
      type: string
      minLength: 1
      maxLength: 64
      examples: [Europe/Zurich]
...
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        trace_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string

    # ---------------------------------------------------------------- auth
...
    SessionStatus:
      type: string
      enum: [draft, completed, abandoned]
    Session:
      type: object
      required: [id, started_at, ended_at, timezone, local_date, title, notes, perceived_fatigue, bodyweight_kg, status, is_rest_day, template_id, completed_at, updated_at, server_updated_at, blocks]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        started_at:
          $ref: "#/components/schemas/Timestamp"
        ended_at:
          $ref: "#/components/schemas/NullableTimestamp"
        timezone:
          $ref: "#/components/schemas/TimeZone"
        local_date:
          type: string
          format: date
        title:
          type: string
        notes:
          type: string
        perceived_fatigue:
          type: [integer, "null"]
        bodyweight_kg:
          type: [number, "null"]
        status:
          $ref: "#/components/schemas/SessionStatus"
        is_rest_day:
          type: boolean
        template_id:
          type: [string, "null"]
          format: uuid
        completed_at:
          $ref: "#/components/schemas/NullableTimestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        server_updated_at:
          $ref: "#/components/schemas/Timestamp"
        blocks:
          description: Live blocks in order, each with its live sets in order.
          type: array
          items:
            $ref: "#/components/schemas/Block"
...
    BlockKind:
      type: string
      enum: [straight, superset, circuit, emom, amrap]
...
    Block:
      type: object
      required: [id, order_index, kind, rounds_planned, rounds_done, interval_s, notes, updated_at, sets]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        order_index:
          type: integer
        kind:
          $ref: "#/components/schemas/BlockKind"
        rounds_planned:
          type: [integer, "null"]
        rounds_done:
          type: [integer, "null"]
        interval_s:
          type: [integer, "null"]
        notes:
          type: string
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        sets:
          type: array
          items:
            $ref: "#/components/schemas/SetEntry"
    SetKind:
      type: string
      enum: [working, warmup, backoff, drop, cluster, test]
...
    SetEntry:
      type: object
      required: [id, block_id, order_index, round_index, kind, is_planned, rest_after_planned_s, rest_after_actual_s, rpe, rir, completed_at, notes, updated_at, elements]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        block_id:
          $ref: "#/components/schemas/Id"
        order_index:
          type: integer
        round_index:
          type: [integer, "null"]
        kind:
          $ref: "#/components/schemas/SetKind"
        is_planned:
          type: boolean
        rest_after_planned_s:
          type: [integer, "null"]
        rest_after_actual_s:
          type: [integer, "null"]
        rpe:
          type: [number, "null"]
        rir:
          type: [integer, "null"]
        completed_at:
          $ref: "#/components/schemas/NullableTimestamp"
        notes:
          type: string
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        elements:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SetElement"
...
    Criteria:
      description: The unlock DSL (brief §6), with defaults made explicit. Empty means self-attest only.
      type: object
      properties:
        all:
          type: array
          items:
            $ref: "#/components/schemas/Condition"
        any:
          type: array
          items:
            $ref: "#/components/schemas/Condition"
    Condition:
      type: object
      required: [exercise, measure, op, value, assistance, occurrences]
      properties:
        exercise:
          $ref: "#/components/schemas/Slug"
        measure:
          type: string
          enum: [reps, hold_seconds, distance_m]
        op:
          type: string
          enum: [">=", ">"]
        value:
          type: number
        assistance:
          type: string
          enum: [none, any]
        min_load_kg:
          type: number
        max_load_kg:
          type: number
        min_form_quality:
          type: integer
        occurrences:
          type: integer
        within_days:
          type: integer
    MapPoint:
      type: object
      required: [constellation, x, y]
      properties:
        constellation:
          type: string
        x:
          type: number
        y:
          type: number
    SkillLevel:
      type: object
      required: [id, slug, name, order, description, est_weeks_from_prev, unlock_criteria, exercises]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        slug:
          $ref: "#/components/schemas/Slug"
        name:
          type: string
        order:
          type: integer
        description:
          type: string
        est_weeks_from_prev:
          description: A display hint shared by everyone; not a promise.
          type: [integer, "null"]
        unlock_criteria:
          $ref: "#/components/schemas/Criteria"
        exercises:
          type: array
          items:
            type: object
            required: [exercise_id, slug, role]
            properties:
              exercise_id:
                $ref: "#/components/schemas/Id"
              slug:
                $ref: "#/components/schemas/Slug"
              role:
                type: string
                enum: [primary_test, progression, accessory, prehab]
    Skill:
      type: object
      required: [id, slug, name, family, difficulty_tier, is_milestone, status, aka, summary, primary_muscles, common_faults, map, levels]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        slug:
          $ref: "#/components/schemas/Slug"
        name:
          type: string
        family:
          $ref: "#/components/schemas/Slug"
        difficulty_tier:
          type: integer
          minimum: 1
          maximum: 10
        is_milestone:
          type: boolean
        status:
          description: "`draft_placeholder` marks content that is not yet researched."
          type: string
          enum: [active, draft_placeholder]
        aka:
          type: array
          items:
            type: string
        summary:
          type: string
        primary_muscles:
          type: array
          items:
            type: string
        common_faults:
          type: array
          items:
            type: string
        map:
          oneOf:
            - $ref: "#/components/schemas/MapPoint"
            - type: "null"
        levels:
          type: array
          items:
            $ref: "#/components/schemas/SkillLevel"
    SkillEdge:
      type: object
      required: [from_level_id, to_level_id, relation, weight]
      properties:
        from_level_id:
          $ref: "#/components/schemas/Id"
        to_level_id:
          $ref: "#/components/schemas/Id"
        relation:
          type: string
          enum: [prerequisite, recommended, alternative, antagonist]
        weight:
          type: number
    SkillGraph:
      type: object
      required: [content_version, skills, edges]
      properties:
        content_version:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/SkillEdge"
    Disclaimer:
      type: object
      required: [code, text]
      properties:
        code:
          type: string
          enum: [educational_only]
        text:
          type: string
    Injury:
      type: object
      required: [region, name, description, risk_factors, early_signs, prehab_exercises, disclaimer]
      properties:
        region:
          type: string
        name:
          type: string
        description:
          type: string
        risk_factors:
          type: array
          items:
            type: string
        early_signs:
          type: array
          items:
            type: string
        prehab_exercises:
          type: array
          items:
            type: object
            required: [exercise_id, slug]
            properties:
              exercise_id:
                $ref: "#/components/schemas/Id"
              slug:
                $ref: "#/components/schemas/Slug"
        disclaimer:
          type: string
          enum: [educational_only]
    SkillDetail:
      allOf:
        - $ref: "#/components/schemas/Skill"
        - type: object
          required: [injuries, injury_disclaimer]
          properties:
            injuries:
              type: array
              items:
                $ref: "#/components/schemas/Injury"
            injury_disclaimer:
              $ref: "#/components/schemas/Disclaimer"
    LevelState:
      type: object
      required: [level_id, state, best_value, best_unit, best_at, first_achieved_at, verification, stale_since, attempts_count]
      properties:
        level_id:
          $ref: "#/components/schemas/Id"
        state:
          type: string
          enum: [locked, available, in_progress, unlocked]
        best_value:
          description: The best logged value for the level's headline condition. Current form, not achievement.
          type: [number, "null"]
        best_unit:
          type: [string, "null"]
          enum: [reps, hold_seconds, distance_m, null]
        best_at:
          $ref: "#/components/schemas/NullableTimestamp"
        first_achieved_at:
          $ref: "#/components/schemas/NullableTimestamp"
        verification:
          type: [string, "null"]
          enum: [auto, self_attested, coach, null]
        stale_since:
          description: A hint that an unlocked skill has not been practised for a while. It never changes `state`.
          $ref: "#/components/schemas/NullableTimestamp"
        attempts_count:
          type: integer
    SkillMap:
      type: object
      required: [content_version, skills, edges, states]
      properties:
        content_version:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/SkillEdge"
        states:
          description: One entry per level of every listed skill.
          type: array
          items:
            $ref: "#/components/schemas/LevelState"
    Unlock:
      type: object
      required: [level_id, skill_slug, skill_name, level_slug, level_name, verification, occurred_at, evidence_set_entry_id]
      properties:
        level_id:
          $ref: "#/components/schemas/Id"
        skill_slug:
          $ref: "#/components/schemas/Slug"
        skill_name:
          type: string
        level_slug:
          $ref: "#/components/schemas/Slug"
        level_name:
          type: string
        verification:
          type: string
          enum: [auto, self_attested, coach]
        occurred_at:
          $ref: "#/components/schemas/Timestamp"
        evidence_set_entry_id:
          type: [string, "null"]
          format: uuid
    XPAward:
      type: object
      required: [source, amount]
      properties:
        source:
          type: string
          enum: [session_completed, skill_unlocked, badge_earned, streak_milestone, plan_adherence, deload_completed]
        amount:
          type: integer
    Streak:
      type: object
      required: [current_days, longest_days, freeze_credits, last_counted_date]
      properties:
        current_days:
          description: Consecutive days trained, rested by plan, or bridged by a freeze.
          type: integer
        longest_days:
          type: integer
        freeze_credits:
          description: Earned automatically; each bridges one missed day.
          type: integer
        last_counted_date:
          type: [string, "null"]
          format: date
    SessionComplete:
      type: object
      additionalProperties: false
      properties:
        completed_at:
          description: When the athlete finished. Defaults to now.
          $ref: "#/components/schemas/Timestamp"
        ended_at:
          $ref: "#/components/schemas/Timestamp"
        perceived_fatigue:
          type: integer
          minimum: 1
          maximum: 10
        bodyweight_kg:
          type: number
          exclusiveMinimum: 0
          maximum: 500
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    CompletionResult:
      type: object
      required: [session, already_completed, unlocked, newly_available, xp_awarded, xp_total, streak]
      properties:
        session:
          $ref: "#/components/schemas/Session"
        already_completed:
          type: boolean
        unlocked:
          description: Levels this completion unlocked, in prerequisite order.
          type: array
          items:
            $ref: "#/components/schemas/Unlock"
        newly_available:
          description: Levels that became available because their prerequisites are now unlocked.
          type: array
          items:
            $ref: "#/components/schemas/Id"
        xp_awarded:
          type: array
          items:
            $ref: "#/components/schemas/XPAward"
        xp_total:
          type: integer
        streak:
          $ref: "#/components/schemas/Streak"
...
    Progress:
      type: object
      required: [xp_total, streak]
      properties:
        xp_total:
          type: integer
        streak:
          $ref: "#/components/schemas/Streak"

    # ----------------------------------------------------------------- sync
...
    MediaUploadRequest:
      type: object
      additionalProperties: false
      required: [id, kind, mime, bytes]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        kind:
          type: string
          enum: [image]
        mime:
          type: string
          enum: [image/jpeg, image/png, image/heic, image/webp]
        bytes:
          description: The exact size of the file to upload.
          type: integer
          format: int64
          minimum: 1
          maximum: 15728640
        width:
          type: integer
          minimum: 1
          maximum: 20000
        height:
          type: integer
          minimum: 1
          maximum: 20000
    Media:
      type: object
      required: [id, kind, mime, bytes, width, height, status, created_at, ready_at, url, url_expires_at]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        kind:
          type: string
          enum: [image, video]
        mime:
          type: string
        bytes:
          type: integer
          format: int64
        width:
          type: [integer, "null"]
        height:
          type: [integer, "null"]
        status:
          type: string
          enum: [pending, ready, failed, quarantined]
        created_at:
          $ref: "#/components/schemas/Timestamp"
        ready_at:
          $ref: "#/components/schemas/NullableTimestamp"
        url:
          description: A short-lived download URL, when the asset is ready.
          type: [string, "null"]
          format: uri
        url_expires_at:
          $ref: "#/components/schemas/NullableTimestamp"
    MediaUpload:
      type: object
      required: [media, upload]
      properties:
        media:
          $ref: "#/components/schemas/Media"
        upload:
          type: object
          required: [method, url, headers, expires_at]
          properties:
            method:
              type: string
              enum: [PUT]
            url:
              type: string
              format: uri
            headers:
              description: Send exactly these headers with the upload; they are part of the signature.
              type: object
              additionalProperties:
                type: string
            expires_at:
              $ref: "#/components/schemas/Timestamp"